Move trie db to ledger
1. rename blockchain to ledger
2. move trie db to blockchain folder
