+++ api/websocket/websocket.go
@@ -7,7 +7,7 @@ import (
 	"github.com/nknorg/nkn/api/common"
 	"github.com/nknorg/nkn/api/websocket/server"
 	. "github.com/nknorg/nkn/block"
-	"github.com/nknorg/nkn/blockchain"
+	"github.com/nknorg/nkn/ledger"
 	. "github.com/nknorg/nkn/common"
 	"github.com/nknorg/nkn/events"
 	"github.com/nknorg/nkn/node"
