+++ consensus/proposing.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	. "github.com/nknorg/nkn/block"
-	"github.com/nknorg/nkn/blockchain"
+	"github.com/nknorg/nkn/ledger"
 	"github.com/nknorg/nkn/util/log"
 	"github.com/nknorg/nkn/util/timer"
 )
