+++ ledger/db/rollback.go
