+++ ledger/stateGenerator.go
@@ -1,8 +1,8 @@
 package blockchain
 
 import (
+	"github.com/nknorg/nkn/ledger/db"
 	"github.com/nknorg/nkn/common"
-	"github.com/nknorg/nkn/db"
 	. "github.com/nknorg/nkn/pb"
 	. "github.com/nknorg/nkn/transaction"
 )
