+++ ledger/trie/encoding.go
