+++ ledger/trie/hasher.go
