+++ ledger/trie/node.go
