+++ ledger/trie/secure_trie.go
