+++ ledger/trie/trie.go
