+++ nknd.go
@@ -16,10 +16,10 @@ import (
 	"github.com/nknorg/nkn/api/httpjson"
 	"github.com/nknorg/nkn/api/httpjson/client"
 	"github.com/nknorg/nkn/api/websocket"
-	"github.com/nknorg/nkn/blockchain"
 	"github.com/nknorg/nkn/consensus"
 	"github.com/nknorg/nkn/crypto"
-	"github.com/nknorg/nkn/db"
+	"github.com/nknorg/nkn/ledger"
+	"github.com/nknorg/nkn/ledger/db"
 	"github.com/nknorg/nkn/node"
 	"github.com/nknorg/nkn/por"
 	"github.com/nknorg/nkn/util/address"
