+++ node/rollback.go
@@ -6,8 +6,8 @@ import (
 	"time"
 
 	. "github.com/nknorg/nkn/block"
-	"github.com/nknorg/nkn/blockchain"
 	"github.com/nknorg/nkn/common"
+	"github.com/nknorg/nkn/ledger"
 	"github.com/nknorg/nkn/util/config"
 	"github.com/nknorg/nnet/log"
 )
